`CTaskComplexFollowPointRoute` (#431)
* CPointRoute: Add useful functions
* CVector:` Normalized` and `Dot`
* CVector2D: `Normalized`
* CVehicle: FIx flag description
* Fix "moveState" in verious classes to be of type eMoveState
* CPointRoute: Improvements
* CTaskComplexWalkRoundCar: IsGoingForDoor
* CTaskSimpleGoTo: Fix flags
* CTaskSimpleGoToPointFine: Ctor default values
* `TaskComplexFollowPointRoute`

+++ source/game_sa/Tasks/TaskTypes/SeekEntity/TaskComplexSeekEntity.h
@@ -105,7 +105,7 @@ public:
             };
         case TASK_COMPLEX_GO_TO_POINT_AND_STAND_STILL:
             return new CTaskComplexGoToPointAndStandStill{
-                (int32)m_moveState,
+                m_moveState,
                 GetSeekPos(ped),
                 m_maxEntityDist2D,
                 m_moveStateRadius,
